+++ R/ms_options.R
@@ -175,7 +175,7 @@ ms_options <- function(...) {
     at_bar_color = NULL,
     at_line_color = NULL,
     at_point_color = "#371ea3",
-    showcaseName = NULL
+    showcase_name = NULL
   )
 
   # input new options